New callout implementation.  This is based on callwheel implementation done by Artur Grabowski and Thomas Nordin for OpenBSD, which is more efficient in several ways than the callwheel implementation that it is replacing.  It has been adapted to our pre-existing callout API, and also provides the slightly more efficient (and much more intuitive) API (adapted to the callout_*() naming scheme) that the OpenBSD version provides.
Among other things, this shaves a bunch of cycles off rescheduling-in-
the-future a callout which is already scheduled, which the common case
for TCP timers (notably REXMT and KEEP).

The API has been simplified a bit, as well.  The (very confusing to
a good many people) "ACTIVE" state for callouts has gone away.  There
is now only "PENDING" (scheduled to fire in the future) and "EXPIRED"
(has fired, and the function called).

Kernel version bump not done; we'll ride the 1.6N bump that happened
with the malloc(9) change.

+++ sys/netiso/tp_pcb.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: tp_pcb.h,v 1.13 2002/05/12 21:30:37 matt Exp $	*/
+/*	$NetBSD: tp_pcb.h,v 1.14 2003/02/04 01:21:05 thorpej Exp $	*/
 
 /*-
  * Copyright (c) 1991, 1993
@@ -261,7 +261,7 @@ struct tp_pcb {
 	int		tp_rtv; /* max round-trip time variance */
 	int		tp_rtt; /* smoothed round-trip time */
 	SeqNum		tp_rttseq;	/* packet being timed */
-	u_int64_t	tp_rttemit;	/* when emitted, in ticks */
+	int		tp_rttemit;	/* when emitted, in ticks */
 	int		tp_idle;/* last activity, in ticks */
 	short		tp_rxtcur;	/* current retransmit value */
 	short		tp_rxtshift;	/* log(2) of rexmt exp. backoff */
